Prevent premature finalisation of an EmpathyPersonaStore in certain cases
In the case that one of the GtkTreeRowReferences belonging to the Personas
in an EmpathyPersonaStore holds the last reference to the persona store, the
store can be prematurely finalised, causing Empathy to go into an infinite
loop inside GHashTable code (the hash table is finalised with the persona
store, and overwritten with 0xfff… — this happens to cause the next bit of
hash table code to be called to go into an infinite loop).

This can be fixed by holding a reference to the persona store when changing
its personas in response to a personas-changed signal.