empathy-message: cast timestamps to glong
tp_message_get_received_timestamp() and tpl_event_get_timestamp() both return
gint64 while the timestamp is stored as a glong. We need to cast it as varargs
won't do it for us and not casting will result in a a sizing mismatch on
32-bit.

Ideally we should stop using time_t and always use gint64 with timestamps.