Adium: Fix selection of default variant
For theme version >= 3, a variant CSS is mandatory. If none are set
inf DefaultVariant, the first one from Variant/ should be used.

For theme version <= 2, a variant CSS is optional, and by default they
should not have one. If no variant is selected, fallback to main.css.

This also prepare needed API to select other variants