ContactWidget: show IRC channel list.
Very recent versions of Idle expose the list of channels provided in
WHOIS replies. The data is provided as a bunch of x-irc-channel fields
with a single value, namely the name of the channel. Here we expose that
data to the user as a comma-separated list of links; clicking a link
ensures that channel.

https://bugzilla.gnome.org/show_bug.cgi?id=592795