[theme-adium] implement edit_message()
This is done via DOM manipulation, we find the id for the <span> tag we
wrapped the message in in the previous commit. We then replace the innerHTML
with the new message.

This requires a version of WebKitGtk 1.3.x, but I'm not precisely sure what. I
used WebKitGtk 1.4.0, but natty only has 1.3.13, so we'll try setting that as
the dep.

Need to somehow mark a message as edited. Could set a class on it, but that
requires support in the theme. Need inspiration here.