PresetDialog: make the window taller
The element we really care about the size of is the list of statuses
past. So I moved the horizontal size request to that—the window has 6px
of padding or so, so this makes it pretty much exactly the same
width—and made the height request be in the golden ratio to it.

I'm probably unusual in having enough statuses to fill my laptop screen
vertically, but, subjectively, this makes the dialog look less stubby
even now I've removed them all to test.

https://bugzilla.gnome.org/show_bug.cgi?id=630943