empathy-video-widget: reimplemented with GtkClutterEmbed
The "min-width" and "min-height" properties are no longer needed as
GtkClutterEmbed uses the space available and works correctly.

With GtkDrawingArea we needed to make all of the initialization in the
->constructed function because we depended on the window's XID and we
could find that only after ->realized was called.

Because we didn't have the XID until at a later time, we could not set
the video sink and connect it to the drawing area until sometime later
when a callback was called.

As we now set the video sink (priv->sink) in ->init (and not in a
callback as we used to with GtkDrawingArea) we don't need several
items:

- the lock: there is nothing to lock, we know the sink from ->init, it
  will not be set from a callback to require locking.

- the FsElementAddedNotifier: this was used to raise a callback when a
  new element was added to the video sink pipeline so we could find
  out which was the X overlay GstElement to connect to the
  GtkDrawingArea's XID.

- the ->sync and ->async booleans were used to set the underlining
  sink's "sync" and "async" properties. Because we might not have had
  a sink when ->set_property or ->get_property were called, we needed
  to keep these until that video sink appeared. Now we simply forward
  these to the underlining sink as it is always known.

CLUTTER_GTK_REQUIRED=1.0.1 is due to a bug that is fixed in
clutter-gtk's head:
  https://bugzilla.gnome.org/show_bug.cgi?id=653435
  embed: unmap: check stage exists before unmapping

Without that patch, after pressing "Hang up", empathy-av (not
Empathy's main window) will SIGSEGV.