PresetDialog: remove support for adding statuses entirely
We already have a way to do this, namely on the contact list itself. So
this dialog now becomes just a way to edit/remove the existing statuses.

The margins on the VBox are chosen to make its right-hand edge line up
exactly with the right-hand edge of the Close button in the dialog.