IRC accounts: top-justify the password hint.
As mentioned on <https://bugzilla.gnome.org/show_bug.cgi?id=654552#c13>,
this makes the password hint text be right up against the field it's
hinting about, at the cost of even more ugly space between it and
the Advanced expander. I think this is the lesser of two evils.