CallWindow: ignore events in the data probe callback
We only want to show the remote video again if we receive
video frames, not if we receive an event for whatever reason.

That's in fact what we get when we start sending our own video,
and so we don't want to display the remote video if it was
hidden because we're not receiving it anymore.