Add a webview widget for displaying the log
The widget is currently bolted in alongside the existing log display widget
for comparison. This includes some debugging cruft like showing the secret
notebook tabs.

The webview is populated from the store_events treestore, allowing all of the
existing node-management and ordering code to be used. Attempting to replace
this logic in Javascript was demonstratably too hard. This approach keeps the
Javascript code down to the 4 GtkTreeModel signals.

TODO:
 - icons
 - pretty printed dates
 - expanders
 - supporting smileys, links, etc.
 - using the correct font, etc.
 - removing the debugging