empathy-call: add a video-effects button
This ads a new button, but it's icon is (at the moment) not good. We
should either add a new icon to the gnome-theme and use that in both
Empathy and Cheese or copy Cheese' icon here.

The button doesn't do anything at the moment.

It is enabled (sensitive) only when we know we have a camera (as
detected by the camera monitor).