empathy-call: add a wrapper box to separate video and effects
We only have the video box now (video_box), but we'll soon be adding
the effects_box.

Separating video and effects in two different boxes makes it easy to
switch from displaying video to displaying effect previews (and
reverse): we only have to hide one of the two boxes to have everything
inside them will be hidden too.