empathy-call: make floating_toolbar visible in effect preview mode
The floating_toolbar was added in the video_layout which is hidden in
effect preview mode.

This patch adds it to the root-bin, making it visible in both
video-mode and effect-preview-mode.

New issue found: clicking inside the floating-toolbar even on buttons
will let the click fall-through to the effect-box bellow and select a
video-preview.