empathy-call: prepare gst pipeline for sending effects on network
Added "ffmpegcolorspace ! identity ! ffmpegcolorspace" to
pipeline. Selecting an effect from the effect-preview page will
replace "identity" with that gstreamer effect.

Also added a "video_preview_valve" to be able to stop sending video on
the network and on the video_preview_sink, but at the same time keep
the video_input in PLAYING so that the effect previews work while we
don't send video on the network.