empathy-call: don't stop the camera while camera preview disabled
We no longer stop the camera while the camera preview is disabled
(when we're not sending video to the other party).

Before this patch there was a bug when calling echo@collabora.co.uk:
 - start a video call with echo@test.collabora.co.uk

 - wait until the camera video is echoed back

 - press the "camera" button to disable camera preview and sending
   video to echo@test.collabora.co.uk

 - wait until the echobot stops transmitting video back

 - press the "camera" button to enable camera preview and sending
   video to echo@test.collabora.co.uk

 - no matter how long you wait video doesn't come back any more.

With this patch, we don't have this problem any more: after pressing
"camera" the second time, we see video streaming back from the echobot
as expected.