libempathy-gtk: Explicitly close the input stream used for reading avatars
Due to a missing unref in GIO, the input stream is leaked by GFileIcon, so
if we rely on its FD being implicitly closed when the stream is finalised,
we'll end up leaking (lots of) file descriptors.

As well as fixing the unref in GIO (see: bgo#657206), we now explicitly
close the input stream so that even if the stream object is leaked, the FD
isn't.

Closes: bgo#656682