Use g_date_time_new_from_unix_local() to read timestamps from logger
g_date_time_new_from_unix_local() reads a Unix timestamp (which is defined
to be in UTC) and sets the timezone for the GDateTime to be the local timezone.
Thus when we format it, we get times in local time.

g_date_time_new_from_unix_utc() is not correct, because it sets the timezone
to UTC.

Fixes: https://bugzilla.gnome.org/show_bug.cgi?id=656189