ContactSelector: mark standard action as the default.
This allows Gtk+ to style the button differently to indicate that it
will be pressed when the user hits Enter; it also fixes hitting enter in
the New Conversation dialog, which does not consider GTK_RESPONSE_ACCEPT
to be a reasonable response and ignores it. So we get a bug fix and
better styling in exchange for less code!

https://bugzilla.gnome.org/show_bug.cgi?id=658343