individual-widget: use GtkGrid instead of GtkTable internally
EmpathyIndividualWidget (used both in the contact information dialogs
and in the tooltips) use a GtkTable internally to layout contact
details.

GtkTable is being deprecated, and does not correctly support
width-for-height size negotiation. So, there are cases where we get
funny overallocations bugs for wrapping GtkLabels; this is especially
annoying for tooltips, as they tend to grow huge vertically.

Porting the code to use a better container such as GtkGrid solves the
bug. This patch does not cover all the use of GtkTable in
EmpathyIndividualWidget, but enough to fix the tooltip sizing.

https://bugzilla.gnome.org/show_bug.cgi?id=658857