Improve iain sleeping pattern
The Hypno Koala was making iain wake up in sweat and with his heart
pounding at unreasonable hours.

Nobody wants that.

The Hypno Koala has served his duty and it's now time for retirement.

Its replacement was designed and drawn by:
  Cyrene Domogalla <cyrene.domogalla@intel.com>

I've included the source SVG in the commit. To generate icons from this
one, one needs to use the render-icon-theme.py script to generate the
16x16, 24x24, 32x32 and 48x48 versions, the 22x22 one was generated by
the generate-png.sh script in data/icons.

https://bugzilla.gnome.org/show_bug.cgi?id=619186