audio-src: Add a caps filter to select appropriate input format
Instead of relying on the default caps that the pipeline selects (which
will usually end up being float32 stereo at 44.1kHz), this sets a caps
filter to select the format we want from pulsesrc -- s16ne mono at 32kHz.

The point of this is to do resampling/conversion as early in the
pipeline as possible, decreasing the amount of data that needs to be
carried around and thus improving performance a bit.