Let the audio source control the volume, not a software volume element
There is no point in amplifying the mic level is it's too low or too
high as it either doesn't have enough information or it is clipping.
Instead tell pulsesrc what it should do, so it can adjust the hardware
volume for us.

Also listen to changed done by pulsesrc so we can track volume updates
and feed back the settings properly