Remove call input volume configuration
As we rely on pulseaudio there is no point in saving our own volume
configuration, pulseaudio does that for us. If we do save it, then we
have no idea which volume to pick after connecting to pulse (the one we
save, the one pulse saved, a random value?).

On top of that our configuration implementation is plain ugly, the
level meter didn't work etc. This is all much better configured using
the gnome-control-center sound panel, which is pretty nice for inputs
and saves its volume on pulse for us to re-use :)