Call: Bind the output properties from the main thread
If we bind and sync the output volume from the non-main thread, we end
up calling gtk and clutter from the non-main thread.. Which leads to
crashes in OpenGL drivers and potenially kills your cat. Instead create
the audio output when the audio content is added, which happens from the
main thread. The output element is still only added to gstreamer
pipeline when needed though.